Improve writes to adapter pool
When the scheduler writes to the adapter pool, it will now spread writes
across the pool starting with the first two and then proceeding to the
next two, and so on. Note the writes will wrap around an uneven list of
adapter clients. For example, for three adapters, the first write will be to
the clients at index 0 and 1. The next write will be to clients at 2,
the last index in the pool, and at 0.

[#139703357]

Signed-off-by: JT Archie <jtarchie@gmail.com>