Use metricemitter to send metrics
Note that as part of this work, we have no nice way to test that the diode
writer in syslog connector does not panic if it ever receives a nil metric
emitter. We've added a feeble test case around the failure, but do not have a
way to assert without injecting the diode writer.

[#144327693]

Signed-off-by: Eno Compton <ecompton@pivotal.io>