Populate LegacySelector to ensure upgrade paths
If the logs-provider does not yet support the new selector pattern,
then it will send everything. This would be an overwhelming amount of
IO for the adapters. Therefore, the adapter will send both the new
and legacy selector to ensure there is some sort of filter.

If the drain type wants both metrics and logs, it will defer to
only logs.