Do not log to tenants first write restries
We are sending to the tenants a message like:

         Syslog Drain: Error when writing. Backing off for 1ms.

every time we retry a connection to a syslog endpoint. But this might
be disturbing and unnecesary as any minor error or connection closed
and reopened would inject the message.

In this commit we only inject this message after the second retry
when the sleep time is significant and it is indeed some problem.

We also log the number of retry.

Fixes https://github.com/cloudfoundry/scalable-syslog/issues/20